% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_function.R
\name{median_function}
\alias{median_function}
\title{A Self-made Median Function}
\usage{
median_function(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to calculate the median from a numeric vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
